requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiomysql==0.2.0
pymysql==1.1.0
redis==5.0.1

# Browser Automation
browser-use==0.1.21
playwright==1.40.0
selenium==4.15.2

# AI and ML
openai>=1.3.8
langchain>=0.0.350
langchain-openai>=0.0.2
beautifulsoup4>=4.12.3
spacy>=3.8.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil==2.8.2

# Async and Background Tasks
celery==5.3.4
kombu==5.3.4

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

